requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
autograd==1.7.0
autoray==0.6.12
Babel==2.14.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.1.4
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.35.42
botocore==1.35.42
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.3
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dm-tree==0.1.8
ecdsa==0.18.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
esptool==4.7.0
executing==2.0.1
facenet-pytorch==2.5.3
fastjsonschema==2.19.1
filelock==3.13.3
filetype==1.2.0
fire==0.7.0
Flask==3.0.2
flask-cors==5.0.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.5.4
gdown==5.1.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.62.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.0
imbalanced-learn==0.12.2
imutils==0.5.4
intelhex==2.3.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.13
kaggle==1.6.14
keras==3.6.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.3.2
mmcv==2.1.0
mmcv-full==1.7.2
mmdet==3.3.0
mmengine==0.10.7
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nlpaug==1.1.11
nltk==3.9.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
PennyLane==0.38.0
PennyLane_Lightning==0.38.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyAudio==0.2.14
pybboxes==0.1.6
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.21
pydantic_core==2.16.3
pyemblib==0.1.2
Pygments==2.17.2
PyMuPDF==1.25.3
pyparsing==3.1.2
pypiwin32==223
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
reedsolo==1.7.0
referencing==0.34.0
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roboflow==1.1.48
rpds-py==0.18.0
rustworkx==0.15.1
s3transfer==0.10.3
safetensors==0.4.3
sahi==0.11.18
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.6.0
Send2Trash==1.8.3
setuptools==68.0.0
shapely==2.0.6
six==1.16.0
smart-open==6.4.0
smbus2==0.4.3
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
split-folders==0.5.1
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
text-unidecode==1.3
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.4.18
timm==1.0.12
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.41.2
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.100
ultralytics-thop==2.0.8
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.25.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wheel==0.43.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yapf==0.43.0
yolov5==7.0.13
